+++ src/solution.rs
@@ -1,8 +1,8 @@
 use std::fmt;
 use std::rc::Rc;
 
-use crate::{ffi, scip_call_panic};
 use crate::variable::Variable;
+use crate::{ffi, scip_call_panic};
 
 /// A wrapper for a SCIP solution.
 pub struct Solution {
@@ -12,17 +12,17 @@ pub struct Solution {
 
 impl Solution {
     /// Returns the objective value of the solution.
-    pub fn get_obj_val(&self) -> f64 {
+    pub fn obj_val(&self) -> f64 {
         unsafe { ffi::SCIPgetSolOrigObj(self.scip_ptr, self.raw) }
     }
 
     /// Returns the value of a variable in the solution.
-    pub fn get_var_val(&self, var: Rc<Variable>) -> f64 {
+    pub fn val(&self, var: Rc<Variable>) -> f64 {
         unsafe { ffi::SCIPgetSolVal(self.scip_ptr, self.raw, var.raw) }
     }
 
     /// Sets the value of a variable in the solution.
-    pub fn set_val(&self, var: &Variable, val: f64) {
+    pub fn set_val(&self, var: Rc<Variable>, val: f64) {
         scip_call_panic!(ffi::SCIPsetSolVal(self.scip_ptr, self.raw, var.raw, val));
     }
 }
@@ -30,7 +30,7 @@ impl Solution {
 impl fmt::Debug for Solution {
     /// Formats the solution for debugging purposes.
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let obj_val = self.get_obj_val();
+        let obj_val = self.obj_val();
         writeln!(f, "Solution with obj val: {obj_val}")?;
         let vars = unsafe { ffi::SCIPgetVars(self.scip_ptr) };
         let n_vars = unsafe { ffi::SCIPgetNVars(self.scip_ptr) };
@@ -47,4 +47,10 @@ impl fmt::Debug for Solution {
         }
         Ok(())
     }
+}
+
+/// Represents and error that can occur when adding a solution.
+pub enum SolError {
+    /// The solution is infeasible.
+    Infeasible
 }